pyproject.toml
[project]
name = "context-builder"
version = "0.1.0"
description = "Context Manager for processing claims documents"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "openai>=1.108.1",
    "python-dotenv>=1.1.1",
    "pytesseract>=0.3.13",
    "pillow>=11.3.0",
    "pypdfium2>=4.30.0",
    "pdf2image>=1.17.0",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.15.1",
]
